use u64 type instead of system-specific value

+++ libvtkrs/include/vtk_object.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "cxx.h"
 
+#include <cstdint>
 #include <vtkObject.h>
 
 void debug_on(vtkObject &obj);
@@ -8,9 +9,9 @@ void debug_off(vtkObject &obj);
 void set_debug(vtkObject &ptr, bool status);
 bool get_debug(const vtkObject &object);
 void modified(const vtkObject &ptr);
-void remove_observer(vtkObject &ptr, unsigned long tag);
-void remove_observers(vtkObject &ptr, unsigned long event);
+void remove_observer(vtkObject &ptr, uint64_t tag);
+void remove_observers(vtkObject &ptr, uint64_t event);
 void remove_all_observers(vtkObject &ptr);
-int has_observer(const vtkObject &ptr, unsigned long event);
+uint64_t has_observer(const vtkObject &ptr, uint64_t event);
 
 void add_observer(vtkObject &obj);
